Разнес фигуры, ограничения и остальную геометрию по разным файлам
